Changed implementation of User feature

+++ src/main/kotlin/com/bashkevich/quizcheckerbackend/services/UserService.kt
@@ -1,35 +1,39 @@
 package com.bashkevich.quizcheckerbackend.services
 
-
+import com.bashkevich.quizcheckerbackend.data.DatabaseFactory.dbQuery
 import com.bashkevich.quizcheckerbackend.data.models.user.CreateUserRequest
 import com.bashkevich.quizcheckerbackend.data.models.user.UpdateUserRequest
 import com.bashkevich.quizcheckerbackend.data.models.user.UserDto
 import com.bashkevich.quizcheckerbackend.data.models.user.toDto
 import com.bashkevich.quizcheckerbackend.data.repositories.UserRepository
+import io.ktor.server.plugins.*
 
 class UserService(private val userRepository: UserRepository) {
 
-    suspend fun getAllUsers(): List<UserDto> {
-        return userRepository.getAllUsers().map { it.toDto() }
+    suspend fun getAllUsers(): List<UserDto> = dbQuery {
+        userRepository.getAllUsers().map { it.toDto() }
     }
 
-    suspend fun getUserById(id: Int): UserDto? {
-        return userRepository.getUserById(id)?.toDto()
+    suspend fun getUserById(id: Int): UserDto? = dbQuery {
+        if (id == 0) throw BadRequestException("Invalid user ID")
+        userRepository.getUserById(id)?.toDto()
     }
 
-    suspend fun createUser(request: CreateUserRequest): UserDto {
+    suspend fun createUser(request: CreateUserRequest): UserDto = dbQuery {
         // Add business logic here (e.g., validation, email verification)
         validateUserRequest(request)
-        return userRepository.createUser(request).toDto()
+        userRepository.createUser(request).toDto()
     }
 
-    suspend fun updateUser(id: Int, request: UpdateUserRequest): Boolean {
+    suspend fun updateUser(id: Int, request: UpdateUserRequest): Boolean = dbQuery {
+        if (id == 0) throw BadRequestException("Invalid user ID")
         // Add business logic here
-        return userRepository.updateUser(id, request)
+        userRepository.updateUser(id, request)
     }
 
-    suspend fun deleteUser(id: Int): Boolean {
-        return userRepository.deleteUser(id)
+    suspend fun deleteUser(id: Int): Boolean = dbQuery {
+        if (id == 0) throw BadRequestException("Invalid user ID")
+        userRepository.deleteUser(id)
     }
 
     private fun validateUserRequest(request: CreateUserRequest) {
